make get_graph private, fix permissions leak
